+++ packages/react-native/src/components/Tags/export.ts
@@ -1,0 +1,2 @@
+export * from "./AlertTab";
+export * from "./RawTag";