+++ packages/react/src/components/Actions/Button/internal.tsx
@@ -1,6 +1,7 @@
 import { Button as ShadcnButton } from "@/ui/button"
 
 import { EmojiImage } from "@/lib/emojis.tsx"
+import { cn } from "@/lib/utils"
 import { cva } from "cva"
 import { ComponentProps, forwardRef, useState } from "react"
 import { Icon, IconType } from "../../Utilities/Icon"
@@ -34,6 +35,7 @@ const iconVariants = cva({
         "text-f1-icon-critical-bold group-hover:text-f1-icon-inverse group-active:text-f1-icon-inverse group-data-[pressed=true]:text-f1-icon-inverse dark:group-hover:text-f1-icon-bold/80 dark:group-active:text-f1-icon-bold/80 dark:group-data-[pressed=true]:text-f1-icon-bold/80",
       ghost: "text-f1-icon",
       promote: "text-f1-icon-promote",
+      outlinePromote: "text-f1-icon-promote",
     },
   },
   defaultVariants: {
@@ -51,7 +53,8 @@ export const iconOnlyVariants = cva({
       critical:
         "text-f1-icon-critical-bold group-hover:text-f1-icon-inverse group-active:text-f1-icon-inverse group-data-[pressed=true]:text-f1-icon-inverse dark:group-hover:text-f1-icon-bold dark:group-active:text-f1-icon-bold dark:group-data-[pressed=true]:text-f1-icon-bold",
       ghost: "text-f1-icon-bold",
-      promote: "text-f1-icon-bold",
+      promote: "text-f1-icon-promote",
+      outlinePromote: "text-f1-icon-promote",
     },
   },
   defaultVariants: {
@@ -126,7 +129,7 @@ const ButtonInternal = forwardRef<HTMLButtonElement, ButtonInternalProps>(
         {emoji && (
           <EmojiImage emoji={emoji} size={size === "sm" ? "sm" : "md"} />
         )}
-        {!hideLabel && label}
+        <span className={cn(hideLabel && "sr-only")}>{label}</span>
         {append}
       </ShadcnButton>
     )