+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.11.0](https://github.com/factorialco/factorial-one/compare/factorial-one-react-native-v0.10.0...factorial-one-react-native-v0.11.0) (2025-06-05)
+
+
+### Features
+
+* create new  Chip & RawTag component for mobile ([#2005](https://github.com/factorialco/factorial-one/issues/2005)) ([3c60e51](https://github.com/factorialco/factorial-one/commit/3c60e5181cbc46db31488dd5755abab45071f008))
+
+## [0.10.0](https://github.com/factorialco/factorial-one/compare/factorial-one-react-native-v0.9.0...factorial-one-react-native-v0.10.0) (2025-06-04)
+
+
+### Features
+
+* create new f1 components AlertTag for Mobile ([#1994](https://github.com/factorialco/factorial-one/issues/1994)) ([afe7df1](https://github.com/factorialco/factorial-one/commit/afe7df11e763712ca5f18e228702c5d09061523f))
+
 ## [0.9.0](https://github.com/factorialco/factorial-one/compare/factorial-one-react-native-v0.8.2...factorial-one-react-native-v0.9.0) (2025-05-30)